feat: Add migration for adding name field to inspections
This commit adds a migration file (AddNameToInspections20240820143000) to the project, which adds a "name" field to the inspections table in the database. The "name" field is of type varchar(255) and has a default value of 'Unnamed Inspection'. This migration enhances the data model by allowing inspections to have a name, improving the clarity and organization of the application's data.